requirements.txt
Flask==2.1.2
torch==1.10.0
torchvision==0.11.1
Pillow==8.4.0
